Tentative fix for unnecessary REGISTER packets.
--------
M      edge.c

Modifed update_peer_address to only call send_register() if the stored IP:port
for the source MAC is different to what is stored.

Also call update_peer_address() with when=time(NULL) instead of when=0.

--------

Commentary:

Before this fix a PACKET type UDP is received in the main edge loop and
update_peer_address() was always called with when=0. This always caused
send_register() to be called, sending an unnecessary REGISTER packet. The new
logic sends a REGISTER to the PACKET sender only if one of these conditions is
satisfied:
 1. The source MAC is unknown.
 2. The stored UDP IP:port for the source MAC is different.