--------
Modified Files

A      edge.c
 B     n2n.c

--------

A: Refactored edge to replace threading.

The use of a second thread is removed as it was found to cause crashes due to
race conditions under heavy load. With these changes edge no longer segfaults
under the load tests which caused it to segfault previously.

select() is now used to determine when input is waiting on either the TAP socket
or the interfacing UDP/TCP socket. The author would choose poll() but it is
unclear if poll() is available under all the target operating systems; while
select() was already in use for the UDP/TCP socket.

B: Remove a compiler warning (unused variable).